\name{grouprar}
\alias{grouprar}
\alias{grouprar-package}
\docType{package}
\title{
  grouprar-package: Group Response-Adaptive Randomization for Clinical Trials
}
\description{
  Implement group response-adaptive randomization procedures, which also integrates standard non-group response-adaptive randomization methods as specialized instances. It is also uniquely capable of managing complex scenarios, including those with delayed and missing responses, thereby expanding its utility in real-world applications. This package offers 16 functions for simulating a variety of response adaptive randomization procedures. These functions are essential for guiding the selection of statistical methods in clinical trials, providing a flexible and effective approach to trial design. Some of the detailed methodologies and algorithms used in this package, please refer to the following references:
    LJ Wei (1979) <doi:10.1214/aos/1176344614>
    L. J. WEI and S. DURHAM (1978) <doi:10.1080/01621459.1978.10480109>
    Durham, S. D., FlournoY, N. AND LI, W. (1998) <doi:10.2307/3315771>
    Ivanova, A., Rosenberger, W. F., Durham, S. D. and Flournoy, N. (2000) <https://www.jstor.org/stable/25053121>
    Bai Z D, Hu F, Shen L. (2002) <doi:10.1006/jmva.2001.1987>
    Ivanova, A. (2003) <doi:10.1007/s001840200220>
    Hu, F., & Zhang, L. X. (2004) <doi:10.1214/aos/1079120137>
    Hu, F., & Rosenberger, W. F. (2006, ISBN:978-0-471-65396-7).
    Zhang, L. X., Chan, W. S., Cheung, S. H., & Hu, F. (2007) <https://www.jstor.org/stable/26432528>
    Zhang, L., & Rosenberger, W. F. (2006) <doi:10.1111/j.1541-0420.2005.00496.x>
    Hu, F., Zhang, L. X., Cheung, S. H., & Chan, W. S. (2008) <doi:10.1002/cjs.5550360404>.
}

\references{
Bai Z D, Hu F, Shen L. \emph{An adaptive design for multi-arm clinical trials}[J]. Journal of Multivariate Analysis, 2002, 81(1): 1-18.

Hu F, Zhang L X. \emph{Asymptotic properties of doubly adaptive biased coin designs for multitreatment clinical trials}[J]. The Annals of Statistics, 2004, 32(1): 268-301.

Hu F, Zhang L X, Cheung S H, et al. \emph{Doubly adaptive biased coin designs with delayed responses}[J]. Canadian Journal of Statistics, 2008, 36(4): 541-559.

Hu F, Rosenberger W F. \emph{Optimality, variability, power: evaluating response-adaptive randomization procedures for treatment comparisons}[J]. Journal of the American Statistical Association, 2003: 671-678.

Ivanova A. \emph{A play-the-winner-type urn design with reduced variability}[J]. Metrika, 2003, 58: 1-13.

Ivanova A, Rosenberger W F, Durham S D, et al. \emph{A birth and death urn for randomized clinical trials: asymptotic methods}[J]. Sankhyā: the Indian Journal of Statistics, Series B, 2000: 104-118

Wei L J. \emph{The generalized Polya's urn design for sequential medical trials}[J]. The Annals of Statistics, 1979, 7(2): 291-296.

Wei L J, Durham S. \emph{The randomized play-the-winner rule in medical trials}[J]. Journal of the American Statistical Association, 1978: 840-843.

Zelen M. \emph{Play the winner rule and the controlled clinical trial}[J]. Journal of the American Statistical Association, 1969: 131-146.

Zhang L X, Chan W S, Cheung S H, et al. \emph{A generalized drop-the-loser urn for clinical trials with delayed responses}[J]. Statistica Sinica, 2007, 17(1): 387-409.

Zhang L, Rosenberger W F. \emph{Response‐adaptive randomization for clinical trials with continuous outcomes}[J]. Biometrics, 2006, 62(2): 562-569.
}

\author{
Guannan Zhai \email{guannanzhai@gwu.edu}; Feifang Hu \email{feifang@gwu.edu}.
}
\keyword{ grouprar-package }
